requirements.txt

flask
flask-cors
mysql-connector-python
pandas
scikit-learn
python-dotenv
